@@ -3,16 +3,16 @@
   "description": "Angular ScrollSpy Service",
   "author": "João Ribeiro <jonnybgod@gmail.com> (http://github.com/JonnyBGod)",
   "devDependencies": {
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
+    "@angular/common": "^6.1.3",
+    "@angular/compiler": "^6.1.3",
+    "@angular/compiler-cli": "^6.1.3",
+    "@angular/core": "^6.1.3",
+    "@angular/platform-browser": "^6.1.3",
+    "@angular/platform-browser-dynamic": "^6.1.3",
+    "@angular/router": "^6.1.3",
     "@types/core-js": "^0.9.41",
     "@types/jasmine": "^2.5.47",
-    "@types/node": "^8.0.47",
+    "@types/node": "^10.1.4",
     "awesome-typescript-loader": "^3.3.0",
     "codelyzer": "^4.0.0",
     "commitizen": "^2.9.6",
@@ -32,16 +32,16 @@
     "ng-packagr": "^1.5.0-rc.1",
     "reflect-metadata": "^0.1.10",
     "rimraf": "^2.6.1",
-    "rxjs": "^5.3.0",
+    "rxjs": "^6.2.2",
     "semantic-release": "^8.2.0",
     "source-map-loader": "^0.2.1",
     "systemjs": "^0.20.12",
     "ts-helpers": "^1.1.1",
     "tslint": "^5.1.0",
     "tslint-loader": "^3.5.2",
-    "typescript": "^2.4.0",
+    "typescript": "^2.9.2",
     "webpack": "^3.8.1",
-    "zone.js": "^0.8.5"
+    "zone.js": "^0.8.26"
   },
   "bugs": {
     "url": "http://github.com/jonnybgod/ngx-scrollspy/issues"
@@ -52,7 +52,7 @@
     "pretest": "npm run rimraf -- dist coverage",
     "test": "karma start",
     "test-watch": "karma start --singleRun=false --autoWatch=true",
-    "clean": "npm cache clean && npm run rimraf -- node_modules dist coverage",
+    "clean": "npm cache clean --force && npm run rimraf -- node_modules dist coverage",
     "preclean:install": "npm run clean",
     "clean:install": "npm set progress=false && npm i",
     "commit": "git-cz",
@@ -85,11 +85,15 @@
   ],
   "license": "MIT",
   "engines": {
-    "node": ">=6.9.0"
+    "node": ">=8.0.0"
   },
   "config": {
     "commitizen": {
       "path": "node_modules/cz-conventional-changelog"
     }
+  },
+  "dependencies": {
+    "@angular/cli": "^6.1.4",
+    "rxjs-tslint": "^0.1.5"
   }
 }
