@@ -16,15 +16,12 @@ require('core-js/es7/reflect');
 require('ts-helpers');
 
 require('zone.js/dist/zone');
+require('zone.js/dist/zone-testing');
 require('zone.js/dist/long-stack-trace-zone');
-require('zone.js/dist/async-test');
-require('zone.js/dist/fake-async-test');
-require('zone.js/dist/sync-test');
 require('zone.js/dist/proxy'); // since zone.js 0.6.15
-require('zone.js/dist/jasmine-patch'); // put here since zone.js 0.6.14
 
 // RxJS
-require('rxjs/Rx');
+require('rxjs');
 
 var testing = require('@angular/core/testing');
 var browser = require('@angular/platform-browser-dynamic/testing');
